+++ src/cli.js
@@ -21,9 +21,7 @@ Options:
 
 import Environment from './environment';
 import Logger from './logger';
-// jshint/jshint#2118
-// import sassdoc, { parse } from './sassdoc';
-import { default as sassdoc, parse } from './sassdoc';
+import sassdoc, { parse } from './sassdoc';
 import * as errors from './errors';
 import { docopt } from 'docopt';
 import source from 'vinyl-source-stream';