+++ src/environment.js
@@ -190,8 +190,6 @@ export default class Environment extends EventEmitter {
    * Process `this.theme`.
    */
   loadTheme() {
-    this.themeName = this.theme || 'default';
-
     if (this.theme === undefined) {
       return this.defaultTheme();
     }
@@ -201,6 +199,7 @@ export default class Environment extends EventEmitter {
     }
 
     let theme = this.resolve(this.theme, this.themeCwd);
+    this.themeName = this.theme;
     this.displayTheme = path.relative(process.cwd(), theme);
 
     return this.tryTheme(theme);
@@ -225,7 +224,7 @@ export default class Environment extends EventEmitter {
 
     if (typeof this.theme !== 'function') {
       this.emit('error', new errors.SassDocError(
-        `Given theme is ${str(this.theme)}, expected ${str(str)}.`
+        `Given theme is ${str(this.theme)}, expected ${str(str)}.` // eslint-disable-line comma-spacing
       ));
 
       return this.defaultTheme();
@@ -251,7 +250,7 @@ export default class Environment extends EventEmitter {
     }
 
     this.theme = require('sassdoc-theme-default');
-    this.themeName = 'default';
+    this.themeName = this.displayTheme = 'default';
   }
 
   /**