+++ test/annotations/see.test.js
@@ -9,9 +9,9 @@ describe('#see', function () {
     assert.deepEqual(see.parse('name'), { type: 'function', name: 'name' });
   });
 
-  it('should rewrite the .toJSON method', function() {
-    var data = [{description: 'desc', context: {name: 'name'}},{see: [see.parse('name')]}];
+  it('should rewrite the .toJSON method', function () {
+    var data = [{ description: 'desc', context: { name: 'name' }}, { see: [see.parse('name')] }];
     see.resolve(data);
-    assert.deepEqual(data[1].see.toJSON(), [ { description: 'desc', context: { name: 'name' } } ]);
+    assert.deepEqual(data[1].see.toJSON(), [{ description: 'desc', context: { name: 'name' }}]);
   });
 });